desync scanning & sending reports

+++ src/scan/debounce/counter.rs
@@ -12,7 +12,7 @@ impl<S, const W: usize, const H: usize> Counter<S, W, H> {
     pub const fn new(scanner: S) -> Self {
         Self {
             scanner,
-            counts: [[DEBOUNCE_COUNT; W]; H],
+            counts: [[DEBOUNCE_COUNT as u16; W]; H],
             states: [[false; W]; H],
         }
     }
@@ -28,12 +28,12 @@ impl<S: Scan<W, H>, const W: usize, const H: usize> Scan<W, H> for Counter<S, W,
                 let cur_state = self.states[y][x];
 
                 if raw_state == cur_state {
-                    self.counts[y][x] = DEBOUNCE_COUNT;
+                    self.counts[y][x] = DEBOUNCE_COUNT as u16;
                 } else {
                     self.counts[y][x] -= 1;
                     if self.counts[y][x] == 0 {
                         self.states[y][x] = raw_state;
-                        self.counts[y][x] = DEBOUNCE_COUNT;
+                        self.counts[y][x] = DEBOUNCE_COUNT as u16;
                     }
                 }
             }
